+++ test/bigtest/constants.js
@@ -1,4 +1,6 @@
 export const loanPolicyId = 'd9cd0bed-1b49-4b5e-a7bd-064b8d177231';
 export const loanPolicyName = 'testLoanPolicyName';
-export const blockedUserId = 'blockedUserId';
+export const blockedUserId1 = 'blockedUserId1';
+export const blockedUserId2 = 'blockedUserId2';
 export const blockedMessage = 'Patron has reached maximum allowed number of items charged out';
+export const manualBlockMessage = 'Invalid email and mailing addresses';