+++ test/bigtest/network/scenarios/manualPatronBlocks.js
@@ -1,10 +1,12 @@
+import { manualBlockMessage } from '../../constants';
+
 export default function patronBlocks(server) {
   server.del('/manualblocks/:id');
 
   server.get('/manualblocks',
     { 'manualblocks': [
       { 'type': 'Manual',
-        'desc': 'Invalid email and mailing addresses.',
+        'desc': manualBlockMessage,
         'staffInformation': 'Last 3 have bounced back and the letter we sent was returned to us.',
         'patronMessage': 'Please contact the Main Library to update your contact information.',
         'expirationDate': '2020-10-23T00:00:00Z',